- Changed time_window_hours from INTEGER to NUMERIC(5,2) in database schema
- Updated OpenAPI spec to use number type with double format
- Added minimum value of 0.1 hours for more precise time windows
- Resolves conflicts and merges time window precision improvements

+++ database/migrations/001_initial_schema.sql
@@ -69,7 +69,7 @@ CREATE TABLE IF NOT EXISTS signal_rules (
     id UUID PRIMARY KEY DEFAULT uuid_generate_v4(),
     name TEXT NOT NULL,
     min_whales INTEGER DEFAULT 3,
-    time_window_hours INTEGER DEFAULT 1,
+    time_window_hours NUMERIC(5, 2) DEFAULT 1.0,
     min_total_sol NUMERIC(20, 9) DEFAULT 10,
     is_active BOOLEAN DEFAULT TRUE,
     created_at TIMESTAMPTZ DEFAULT NOW()
